Force unlink to check for non-null path
Context: pathname argument in the unlink function is supposed to
be non-null. But the function can still be called with a NULL value.
This is causing an issue with the IME native API.

This patch forces the unlink function to test the pathname argument.
Do do so the function is compiled with O0 flag.